Create CatYearsDogYears.js
Cat Years Dog Years - Use the variable humanYears to calculate the cat and dogs their respective ages in cat years and dog years. For the cat, +15 for first year, +9 for second year and +4 for each year after. For the dog, +15 for first year, +9 for second year and +5 for each year after.